Workaround for a problem when compiling many files on Windows.
I've modified Matt's patch a little to kick in only if we are actually
running on Windows and want to compile a lot of files.

Unfortunately File.createTempFile is not available on JDK 1.1. I use
java.util.Random to create a hopefully unique filename for a temporary
file. Hope this works - don't have a Windows box to check.

Submitted by:	mpfoemme@ThoughtWorks.com


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@267810 13f79535-47bb-0310-9956-ffa450edef68