Great deal of modification around style tasks... Mostly I wanted to unify the syntax and deal with systemid but it unfortunately leads to problems because the URI specs are not clearly defined and xsl processors do what they want.... Inconsistencies in Xalan 2 prevented the use of File objects (bugs are currently being resolved), thus I switched to systemid that are known to work fine and should work fine in later versions.
Fixed a bug in XslpLiaison that was producing a NPE.
(did someone use that task before ?)

Added testcases that are transforming an XML file via an XSL.
The point here is that the XML and the XSL do an include, this allow to check
that the resolution is working fine and thus the systemid is correctly setted.
Tescases also check for redirect behavior (xalan extension) end do a
mini-encoding test (an encoding attribute should probably be added as a
convenience to the style task)

Added Javadoc as well in the XSLTLiaison interface.

Related to bug 3087, 3088.


git-svn-id: https://svn.apache.org/repos/asf/ant/core/branches/ANT_14_BRANCH@269548 13f79535-47bb-0310-9956-ffa450edef68