Fixed systemid bug that shows up only with crimson (bug #3132)
The TraXLiaison testcase was demonstrating this error but it didn't show up
because I had xerces in the path and xerces handles this without complaining.

Crimson expect systemid as being fully /'ed. It gets the base uri by
doing a lastIndexOf('/') meaning that windows like pathname were
returning a file:/// as the base uri...

Now the systemid is /'ed which should be the standard...


git-svn-id: https://svn.apache.org/repos/asf/ant/core/branches/ANT_14_BRANCH@269571 13f79535-47bb-0310-9956-ffa450edef68