The attached patch makes <path> useable as a data-type in myrmidon.
* Added a String -> Path converter.

* Added the appropriate declarations to ant1-ant-descriptor.xml.

* PathLocation is now only used internally by Path.  This means a <path> may
not contain nested <pathlocation> elements any more.  Nested <path> elements
can be used to do the same thing.

* Removed Path.systemClasspath and Path.concatSystemClassPath().  The goal
is to add specialised <systemclasspath>, <antruntime>, and <javaruntime>
data-types to control this explicitly.  I left it unfinished, because the
as-yet-unwritten Java util stuff will determine how it should be done.

* Moved Path.addExtdirs() -> DefaultCompilerAdaptor.  This was the only
place it was used.

* Cleaned out a few more Path createX() methods.


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@270786 13f79535-47bb-0310-9956-ffa450edef68