This patch adds bunch of minor features to the configurer:
* Added max multiplicity checking.  Properties with a setter method can only
be set once, whereas properties with an adder method can be set an unlimited
number of times.

* Resolves properties in reference ids.  e.g

  <javac classpath-ref="${my-classpath-id-name}"/>

* Ignores String adder and setter methods, if other methods exist.  Longer
term, the type should be able to specify exactly which method to use.

* Moved all per-object state behind the ConfigurationState interface.  The
ObjectConfigurer is now responsible for state-based validation.

* Tidied-up error messages.  More context info is available in error
messages, to make figuring out the problem easier.  Error messages still
need work.


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@270822 13f79535-47bb-0310-9956-ffa450edef68