Quick change to ProjectHelperImpl to use SAX2.
Note that it already had dependencies on SAX2 interfaces, but the original
helper worked with SAX1 parsers too. This one will only work with SAX2.

IMHO SAX2 should be the default, and the old one should be kept
around for backward compat or special cases.

The parser is not using the namespace in any way - it'll use the
RoleFactory plugin to allow use of the namespace in task creation.


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@271736 13f79535-47bb-0310-9956-ffa450edef68