cvschangelog detects ends of comments by looking for a line starting with ------
Unfortunately if somebody commits a file using a message like:
"added a new feature
--------
removed an old feature"

The second line will be identified as the end of the comment. The same for the
end of a changelog with '======'.

Using the right number of '-' or '=' fixes the problem temporarly until one
commit a file with a number of '-' or '=' greater or equals to the cvs
separator.

PR: 8085
Submitted by: fred@castify.net (Frederic Lavigne)


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@272427 13f79535-47bb-0310-9956-ffa450edef68