Small update to the SAX2 reader - use a class that extends RuntimeConfigurable and supports SAX2 attributes.
In addition, the property substitution has few enhancements: if the
property name is not found and it has a special syntax we'll use the
references to compute a value.

This is work in progress, but the intention is to support dom: syntax
for selecting nodes ( similar with XmlProperties, but using 'life'
DOM tree ) and bean: to use getters on the tasks/objects in the reference
table.


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@272726 13f79535-47bb-0310-9956-ffa450edef68