Why this has not been done before is beyond me, but its something I can't live without!  :)
I typically have properties that refer to a JAR file, but its very difficult to construct a fileset to copy it or include it in a WAR/EAR - so this change fixes this.  <fileset file="${some.jar}"/>

If this change is ok with everyone I'll take care of documenting it.


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@273222 13f79535-47bb-0310-9956-ffa450edef68