Few fixes ( I hope it won't brake anything ).
Added 'org.sax' and 'sun.reflect' to the list of system pacakges. There are
problems when using JDK1.4 with 'endorsed' not set up corectly. This allows
using a different parser, but doesn't allow the overriting of system classes.

Made loadClass synchronized. Yes, it must be - there is a race condition,
we check if the class was not loaded and then load it, but 2 threads
may end up loading the class, and the second will get an exception.
This happen when using tomcat in ant for example - probably other
multithreaded tasks as well.

I left the stack trace in, it shouldn't happen in normal cases ( AFAIK ),
but when something strange happens it helps a lot to know what was wrong.


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@273239 13f79535-47bb-0310-9956-ffa450edef68