Added the proposed hook for component creation.
Full backward compatibility is provided - the project
is used to create tasks. The mechanism allows full
flexibility in implementing any kind of antlib policy,
and allows applications embeding ant to integrate in the
component creation process.

Probably a good improvement would be to move the 'default'
behavior in a separate plugin ( DefaultComponentFactory - to
use the project component storage ), and deprecate the use
of Project for project component management.

After more feedback is received I'll propose this for the
main branch.


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@273430 13f79535-47bb-0310-9956-ffa450edef68