Merged the patch from Nicola.
Removed some stuff that was not used.

If the dynamic properties are adopted in the main tree, we
can remove ( or move out ) the optional implementations.

I also removed the jmx stuff - there is no need to have them
tied to embed or ant1.6. Some functional ( but not complete )
jmx tasks are already in jakarta-commons/modeler ( including
code to control tomcat5 using jmx )


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@273573 13f79535-47bb-0310-9956-ffa450edef68