I hope this won't brake anything :-), but I need to try it.
Remove the reference to the task after the task execution.
Next time the task is called a new instance will be created.

This avoids excessive memory usage and potential leaks in large
programs. All the memory that is used by the task - i.e. the
task itself and all the object it creates - are beeing hold in
memory and prevented from GC by this reference.

I hope this will remove some OutOfMemory errors for large build
files.

Nicola - please let me know if you see any problems after this change !


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@273574 13f79535-47bb-0310-9956-ffa450edef68