Added a hashtable that allows delayed initialization.
In most build files you don't use _all_ tasks and types -
and it takes few second to process all exceptions due to class not
found ( for optional tasks ). This class will allow tasks to
be created when needed, and will also improve the behavior
when the main class loader is changed.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@273619 13f79535-47bb-0310-9956-ffa450edef68