2 small (?) changes.
References are stored in a special hashtable that automatically
configures UnknownElements. ( this will probably be a noop with
the current execution model most of the times - since ProjectHelper
already does that ).

Also task and type creation is delayed and not all tasks are
constructed. That should remove few seconds from the startup time,
and will simplify a lot of code that modifies the classpath at
runtime.

Please review - and if you see any problem -1 it. I checked it in
mostly to allow other people to verify the Script with lazy
eval. If there are objections I can turn this into a hook
and move it to [embed].

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@273620 13f79535-47bb-0310-9956-ffa450edef68