Add a cloneVm attribute to CommandlineJava and <java> and <junit>.
If set, the forked VM will be configured to match the currently
running VM closely.  This involves:

* copying of all system properties.
* copying of the bootclasspath - only if no bootclasspath has been
  specified explicitly or build.sysclasspath has been set to "only".

This is accompanied by a magic system property build.clonevm that can
be used to force the attribute to be set.

It has to be a system property as CommandlineJava doesn't know about
project instances.

PR: 25327


git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@275775 13f79535-47bb-0310-9956-ffa450edef68